Rename exercise render files and route components
