+++ apps/www/app/[locale]/(app)/(shared)/try-out/[product]/[slug]/part/[partKey]/page.tsx
@@ -14,8 +14,8 @@ import { Effect } from "effect";
 import { notFound, redirect } from "next/navigation";
 import { hasLocale } from "next-intl";
 import { getTranslations } from "next-intl/server";
-import { ExerciseArticle } from "@/components/exercise/render/article";
-import { QuestionAnalytics } from "@/components/exercise/render/question-analytics";
+import { QuestionAnalytics } from "@/components/exercise/item/analytics";
+import { ExerciseArticle } from "@/components/exercise/item/article";
 import { TryoutPartRouteShell } from "@/components/tryout/part-route-shell";
 import { TryoutPartRuntime } from "@/components/tryout/part-runtime";
 import { TryoutPartProvider } from "@/components/tryout/providers/part-provider";
